Create Encrypter
We are taking a plain text image bringing out the RGB from it and using the XOR function to encrypt it into something totally different. This method is the very popular One-time Pad XOR cipher.
Notice that the plaintext only shines through 50% of the time, which results in noise as each pixel is equally likely to be 0 or 1.
This image contains no information about the original image. If we didn’t provide the shift sequence it would be impossible for you to reverse it back to the original image. You could try every possible sequence, but that would result in every possible image! How could you know it was the original? It's equally likely to be a picture of you or anything else you can think of.